Mode selection now working w/ connect problems
Both normal and wave modes work. During the connect process when
the python program tries to send the mode select string, it often
takes many attempts to get the mode selected properly. I'm not sure
if it's an issue with the FTDI chip or some issue between python
and the Arduino's serial library.